Move getSocialProfile and getSocialContent to lib/social-platforms
