+++ apps/web/lib/bounty/api/get-social-metrics-updates.ts
@@ -1,5 +1,5 @@
-import { getSocialContent } from "@/lib/api/scrape-creators/get-social-content";
 import { resolveBountyDetails } from "@/lib/bounty/utils";
+import { getSocialContent } from "@/lib/social-platforms/get-social-content";
 import { Bounty, BountySubmission } from "@dub/prisma/client";
 
 export type SocialMetricsUpdate = Pick<
